Get rid of controller interface.
We've considered supporting controller interface.
But it turned out it's unnecessary here.
It just makes ui_viewmgr complex and we couldn't find out the actual usages.

If it's usage is figured out, then we could revert this interface again.

But at least, users could inherit ui_view and override or implement behaviors
from it's interfaces.

Change-Id: Ie58ce01aaee8708371e2ed7e187a7910624cb6ea